feat: add describe_indices function (#5221)
The list_indices function relies on APIs from the index objects
themselves. This means we need to load the indices to populate the
information. In addition, the python function uses the index statistics
which can be slow.

Rather than modify the existing method (which may introduce a breaking
change) this creates a new method `describe_indices`. This method only
uses information available in the dataset manifest. This ensures that
minimal I/O will be required (loading the manifest if it hasn't been
loaded) and the call shouldn't be slow.