feat: support credentials vending for file reader and session (#5256)
Allow user to directly create a LanceFileReader or LanceFileSession with
storage options provider to refresh vended credentials.

We also add a few methods in LanceFileSession for generic object store
usage, these are more as workarounds because there is no practical
solutions today for an IO library with pluggable credentials provider.
The long term plan is to use OpenDAL for related workloads.