fix: handle logical rows deletion properly for zonemap and bloomfilter (#5140)
The old zonemap && bloomfilter use the logical row concept to define
zones. It breaks when the rows are not contiguous, e.g. deletion. This
PR now ensures that the fragment offset is taken into consideration so
that the new "physical row address" handles deletion scenario.

Both Rust and Python tests are added.

close #4758

---------

Co-authored-by: Will Jones <willjones127@gmail.com>