refactor: separate out python and java LanceNamespace interface (#5364)
This is a partial revert of #5345. Because lancedb does not depend on
pylance (except for in tests), we cannot put the LanceNamespace
pluggable interface in pylance directly, so move back to use
lance_namesapce python module. Also do the same for java for
consistency. For rust, because it is a source build, we have concluded
that it should remain in core lance repo for easy import.

For more details, see latest
https://lance.org/community/project-specific/namespace/#dependency-diagram