fix: add graceful shutdown and start for rest namespace adapter (#5325)
Closes #5293 

The test was flaky because the server in the last test is not properly
shut down, and the one in new test fails to start and then client talks
to the server in the last test.

Made the RestAdapter run non-blocking start instead of blocking serve,
so that the python and java level can more gracefully handle start and
shutdown.