refactor: use the same path for dedicated and packed blob (#5449)
I made this change to make it easier for us to perform compaction or GC,
as all blob IDs will now refer to the same blob paths. This means that
as long as we know the largest blob IDs, we can simply remove them all
at once.

---

**Parts of this PR were drafted with assistance from Codex (with
`gpt-5.1-codex-max`) and fully reviewed and edited by me. I take full
responsibility for all changes.**

Signed-off-by: Xuanwo <github@xuanwo.io>