ci: update tpc-h dataset generation to use less ram (#5477)
The TPC-H dataset generation currently materializes the table in RAM at
least twice. Since this table is close to 10GB it can require a lot of
memory and our CI benchmarking jobs are failing.

This PR changes the dataset generation to use disk and streaming so that
it uses much less RAM. There is little change to the overall data
generation time.