refactor: support java 21, drop java 8 (#5565)
Drop the Java 8 build from CI. Java 8 is more than 10 years old, and has
been end-of-life (EOL) since 2019. It is no longer receiving security
updates and other fixes upstream.

Change the release to build with Java 11, since this is now the lowest
Java version we support.

Add a Java 21 build to CI. Java 21 was released more than 2 years ago
and is an LTS release that will probably be used for some time to come.

In the Java code, upgrade Apache Arrow from version 15.0.0 to version
18.3.0. The new version supports Java 21, and does not support Java 8.