ci: make sure changes in lance-namespace-impls doc is auto-updated (#5569)
In previous change, we manually added each supported catalog
integration, but that means every time we add a new one we still need to
make a PR to lance, and that is inconvenient. This PR makes it simpler
by having the nav `.pages` in the `lance-namespace-impls` repo and just
add the template in the end, so we only need to update that repo for new
catalog implementation supports.