feat: support credentials vending in directory namespace (#5566)
This PR introduces the credentials vending feature to the namespace
impl, allowing us to vend credentials if we run directory namespace, or
run it as backend for rest namespace. This would allow us to fully test
the credentials vending code path end to end.

The actual vending logic mainly consults the same feature implemented in
Apache Polaris. The support covers aws, gcp and azure.