feat: upgrade lance-namespace to 0.4.0 (#5568)
Make fixes for the following breaking changes in 0.4.0:
1. introduced full error handling spec, update rust interface to
implement the spec, and also dir and rest implementations in rust,
python, java based on it
2. fix that `create_empty_table` is deprecated and `declare_table` is
introduced. We will start to use `declare_table`, and mark
`create_empty_table` as deprecated and will be removed in 3.0.0. If
`declare_table` fails, it currently falls back to `create_empty_table`.
3. we add `deregister_table` support for dir namespace (without
manifest), by adding a `.lance-deregistered` file. So when checking
table existence, we check (1) the table directory exists, (2) if table
directory exists, the `.lance-deregistered` file does not exist. This
allows user to deregister a table immediately without deleting data
which could be long running.