perf: reuse zstd compressors in encoding (#5598)
Prior to this commit, the ZstdBufferCompressor would construct a new
zstd stream encoder on every call to compress. With this change we
create one compression context for the ZstdBufferCompressor, and reuse
it across calls to compress.

Reuse is not implemented for lz4 compression or for decompression. These
were both explored but did not bring meaningful benefits over the
existing code.